+++ web/app/pages/Project/View/utils/filters.tsx
@@ -6,8 +6,8 @@ import _some from 'lodash/some'
 import _startsWith from 'lodash/startsWith'
 import React from 'react'
 
+import { ProjectPreferences } from '../../providers/CurrentProjectProvider'
 import { Filter } from '../interfaces/traffic'
-import { ProjectPreferences } from '../providers/CurrentProjectProvider'
 
 export const ERROR_FILTERS_MAPPING = {
   showResolved: 'showResolved',
