+++ web/app/routes/projects.$id.tsx
@@ -3,8 +3,8 @@ import { type LinksFunction, type LoaderFunctionArgs, type MetaFunction } from '
 
 import { useRequiredParams } from '~/hooks/useRequiredParams'
 import { API_URL } from '~/lib/constants'
+import { CurrentProjectProvider } from '~/pages/Project/providers/CurrentProjectProvider'
 import ViewProject from '~/pages/Project/View'
-import { CurrentProjectProvider } from '~/pages/Project/View/providers/CurrentProjectProvider'
 import ProjectViewStyle from '~/styles/ProjectViewStyle.css?url'
 import { detectTheme, isEmbedded, isAuthenticated, getProjectTabs } from '~/utils/server'
